TASK: Refactor and unify tests
