+++ Tests/Functional/Traits/ContentRepositoryMultiDimensionNodeCreationTrait.php
@@ -13,66 +13,10 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Tests\Functional\Trait
  * source code.
  */
 
-use Flowpack\ElasticSearch\Transfer\Exception\ApiException;
-use Neos\ContentRepository\Domain\Model\NodeInterface;
-use Neos\ContentRepository\Domain\Model\Workspace;
-use Neos\ContentRepository\Domain\Repository\NodeDataRepository;
-use Neos\ContentRepository\Domain\Repository\WorkspaceRepository;
-use Neos\ContentRepository\Domain\Service\Context;
-use Neos\ContentRepository\Domain\Service\ContextFactoryInterface;
-use Neos\ContentRepository\Domain\Service\NodeTypeManager;
-use Neos\ContentRepository\Exception\NodeExistsException;
-use Neos\ContentRepository\Exception\NodeTypeNotFoundException;
-use Neos\Flow\Cli\Exception\StopCommandException;
-use Neos\Flow\Mvc\Exception\StopActionException;
-
 trait ContentRepositoryMultiDimensionNodeCreationTrait
 {
     /**
-     * @var WorkspaceRepository
-     */
-    protected $workspaceRepository;
-
-    /**
-     * @var NodeInterface
-     */
-    protected $siteNode;
-
-    /**
-     * @var ContextFactoryInterface
-     */
-    protected $contextFactory;
-
-    /**
-     * @var NodeTypeManager
-     */
-    protected $nodeTypeManager;
-
-    /**
-     * @var NodeDataRepository
-     */
-    protected $nodeDataRepository;
-
-    private function setupContentRepository():void
-    {
-        $this->workspaceRepository = $this->objectManager->get(WorkspaceRepository::class);
-        $liveWorkspace = new Workspace('live');
-        $this->workspaceRepository->add($liveWorkspace);
-
-        $this->nodeTypeManager = $this->objectManager->get(NodeTypeManager::class);
-        $this->contextFactory = $this->objectManager->get(ContextFactoryInterface::class);
-        $this->nodeDataRepository = $this->objectManager->get(NodeDataRepository::class);
-    }
-
-    /**
      * Creates some sample nodes to run tests against
-     *
-     * @throws NodeExistsException
-     * @throws NodeTypeNotFoundException
-     * @throws StopActionException
-     * @throws \Flowpack\ElasticSearch\ContentRepositoryAdaptor\Exception
-     * @throws ApiException
-     * @throws StopCommandException
      */
     protected function createNodesForNodeSearchTest(): void
     {
