+++ Tests/Functional/Traits/ContentRepositoryNodeCreationTrait.php
@@ -13,59 +13,14 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Tests\Functional\Trait
  * source code.
  */
 
-use Flowpack\ElasticSearch\Transfer\Exception\ApiException;
-use Neos\ContentRepository\Domain\Model\NodeInterface;
-use Neos\ContentRepository\Domain\Model\Workspace;
-use Neos\ContentRepository\Domain\Repository\NodeDataRepository;
-use Neos\ContentRepository\Domain\Repository\WorkspaceRepository;
-use Neos\ContentRepository\Domain\Service\Context;
-use Neos\ContentRepository\Domain\Service\ContextFactoryInterface;
-use Neos\ContentRepository\Domain\Service\NodeTypeManager;
-use Neos\ContentRepository\Exception\NodeExistsException;
-use Neos\ContentRepository\Exception\NodeTypeNotFoundException;
-use Neos\Flow\Cli\Exception\StopCommandException;
-use Neos\Flow\Mvc\Exception\StopActionException;
-
 trait ContentRepositoryNodeCreationTrait
 {
-    /**
-     * @var WorkspaceRepository
-     */
-    protected $workspaceRepository;
 
     /**
-     * @var Context
+     * Creates some sample nodes to run tests against
      */
-    protected $context;
-
-    /**
-     * @var NodeInterface
-     */
-    protected $siteNode;
-
-    /**
-     * @var ContextFactoryInterface
-     */
-    protected $contextFactory;
-
-    /**
-     * @var NodeTypeManager
-     */
-    protected $nodeTypeManager;
-
-    /**
-     * @var NodeDataRepository
-     */
-    protected $nodeDataRepository;
-
-    private function setupContentRepository():void
+    protected function createNodesForNodeSearchTest(): void
     {
-        $this->workspaceRepository = $this->objectManager->get(WorkspaceRepository::class);
-        $liveWorkspace = new Workspace('live');
-        $this->workspaceRepository->add($liveWorkspace);
-
-        $this->nodeTypeManager = $this->objectManager->get(NodeTypeManager::class);
-        $this->contextFactory = $this->objectManager->get(ContextFactoryInterface::class);
         $this->context = $this->contextFactory->create([
             'workspaceName' => 'live',
             'dimensions' => ['language' => ['en_US']],
@@ -76,21 +31,6 @@ trait ContentRepositoryNodeCreationTrait
         $this->siteNode = $rootNode->createNode('welcome', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
         $this->siteNode->setProperty('title', 'welcome');
 
-        $this->nodeDataRepository = $this->objectManager->get(NodeDataRepository::class);
-    }
-
-    /**
-     * Creates some sample nodes to run tests against
-     *
-     * @throws NodeExistsException
-     * @throws NodeTypeNotFoundException
-     * @throws StopActionException
-     * @throws \Flowpack\ElasticSearch\ContentRepositoryAdaptor\Exception
-     * @throws ApiException
-     * @throws StopCommandException
-     */
-    protected function createNodesForNodeSearchTest(): void
-    {
         $newDocumentNode1 = $this->siteNode->createNode('test-node-1', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
         $newDocumentNode1->setProperty('title', 'chicken');
         $newDocumentNode1->setProperty('title_analyzed', 'chicken');
