+++ Tests/Functional/Traits/ContentRepositorySetupTrait.php
@@ -1,0 +1,66 @@
+<?php
+declare(strict_types=1);
+
+namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Tests\Functional\Traits;
+
+/*
+ * This file is part of the Flowpack.ElasticSearch.ContentRepositoryAdaptor package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+
+use Neos\ContentRepository\Domain\Model\NodeInterface;
+use Neos\ContentRepository\Domain\Model\Workspace;
+use Neos\ContentRepository\Domain\Repository\NodeDataRepository;
+use Neos\ContentRepository\Domain\Repository\WorkspaceRepository;
+use Neos\ContentRepository\Domain\Service\Context;
+use Neos\ContentRepository\Domain\Service\ContextFactoryInterface;
+use Neos\ContentRepository\Domain\Service\NodeTypeManager;
+
+trait ContentRepositorySetupTrait
+{
+    /**
+     * @var WorkspaceRepository
+     */
+    protected $workspaceRepository;
+
+    /**
+     * @var Context
+     */
+    protected $context;
+
+    /**
+     * @var NodeInterface
+     */
+    protected $siteNode;
+
+    /**
+     * @var ContextFactoryInterface
+     */
+    protected $contextFactory;
+
+    /**
+     * @var NodeTypeManager
+     */
+    protected $nodeTypeManager;
+
+    /**
+     * @var NodeDataRepository
+     */
+    protected $nodeDataRepository;
+
+    private function setupContentRepository():void
+    {
+        $this->workspaceRepository = $this->objectManager->get(WorkspaceRepository::class);
+        $liveWorkspace = new Workspace('live');
+        $this->workspaceRepository->add($liveWorkspace);
+
+        $this->nodeTypeManager = $this->objectManager->get(NodeTypeManager::class);
+        $this->contextFactory = $this->objectManager->get(ContextFactoryInterface::class);
+        $this->nodeDataRepository = $this->objectManager->get(NodeDataRepository::class);
+    }
+}
